Funcao fitness

+++ src/main/java/br/usp/ia/logging/Logging.java
@@ -1,8 +1,8 @@
 package br.usp.ia.logging;
 
 public interface Logging {
-    void fitnessProgress(int generation, int generationScreenSkip, double populationTotal, double populationAvg,
-                         double max, double min);
+    void fitnessProgress(int generation, int generationScreenSkip, Integer populationTotal, double populationAvg,
+                         Integer max, Integer min);
 
     void print(String string);
 }
